+++ src/components/footer.tsx
@@ -1,0 +1,7 @@
+export default function Footer(){
+    return(
+        <footer className="mt-6 text-gray-500 text-sm">
+            © 2025 Mi Aplicación de Economía
+        </footer>
+    )
+}
