+++ src/components/header.tsx
@@ -1,0 +1,8 @@
+export default function Header(){
+    return(
+    <header className="bg-white shadow-md rounded-lg w-full max-w-xl p-6 mb-6">
+        <h1 className="text-3xl font-bold text-gray-800 text-center">
+            Mi Aplicación de Economía
+        </h1>
+    </header>)
+}
