+++ src/features/Transacciones/types/transaccion.d.ts
